requirements.txt
absl-py                  
asttokens                
backcall                 
blas                     
bottleneck               
brotli-python            
bzip2                    
c-ares                   
ca-certificates          
cairo                    
certifi                  
charset-normalizer       
colorama                 
comm                     
contourpy                
cuda-cccl                
cuda-cccl_win-64         
cuda-cudart              
cuda-cudart-dev          
cuda-cupti               
cuda-libraries           
cuda-libraries-dev       
cuda-nvrtc               
cuda-nvrtc-dev           
cuda-nvtx                
cuda-opencl              
cuda-opencl-dev          
cuda-profiler-api        
cuda-runtime             
cuda-version             
cycler                   
debugpy                  
decorator                
eigen                    
exceptiongroup           
executing                
expat                    
fightingcv-attention     
filelock                 
fire                     
fontconfig               
fonttools                
freetype                 
fsspec                   
git                      
glib                     
glib-tools               
gmp                      
gmpy2                    
graphite2                
grpcio                   
gst-plugins-base         
gstreamer                
harfbuzz                 
hdf5                     
icc_rt                   
icu                      
idna                     
importlib-metadata       
importlib_metadata       
importlib_resources      
intel-openmp             
ipdb                     
ipykernel                
ipython                  
jedi                     
jinja2                   
joblib                   
jpeg                     
jsonpatch                
jsonpointer              
jupyter_client           
jupyter_core             
khronos-opencl-icd-loader
kiwisolver               
krb5                     
lcms2                    
lerc                     
libabseil                
libclang                 
libclang13               
libcublas                
libcublas-dev            
libcufft                 
libcufft-dev             
libcurand                
libcurand-dev            
libcusolver              
libcusolver-dev          
libcusparse              
libcusparse-dev          
libdeflate               
libffi                   
libglib                  
libgrpc                  
libiconv                 
libjpeg-turbo            
libnpp                   
libnpp-dev               
libnvjitlink             
libnvjitlink-dev         
libnvjpeg                
libnvjpeg-dev            
libogg                   
libpng                   
libpq                    
libprotobuf              
libsodium                
libtiff                  
libuv                    
libvorbis                
libwebp-base             
libxml2                  
lz4-c                    
markdown                 
markupsafe               
matplotlib               
matplotlib-base          
matplotlib-inline        
mkl                      
mkl-service              
mkl_fft                  
mkl_random               
mpc                      
mpfr                     
mpmath                   
mypy-extensions          
nest-asyncio             
networkx                 
numexpr                  
numpy                    
opencv                   
openjpeg                 
openssl                  
packaging                
pandas                   
parso                    
pcre2                    
pickleshare              
pillow                   
pip                      
pixman                   
platformdirs             
ply                      
prettytable              
prompt-toolkit           
protobuf                 
psutil                   
pure_eval                
pybind11-abi             
pygments                 
pyparsing                
pyqt                     
pyqt5-sip                
pyqt6-sip                
pyre-extensions          
pysocks                  
python                   
python-dateutil          
python-tzdata            
pytorch                  
pytorch-cuda             
pytorch-mutex            
pytz                     
pywin32                  
pyyaml                   
pyzmq                    
qt-main                  
re2                      
requests                 
scikit-learn             
scipy                    
seaborn                  
setuptools               
sip                      
six                      
sqlite                   
stack_data               
sympy                    
tabulate                 
tbb                      
tensorboard              
tensorboard-data-server  
termcolor                
threadpoolctl            
tomli                    
torchaudio               
torchnet                 
torchsummary             
torchtnt                 
torchvision              
tornado                  
tqdm                     
traitlets                
typing-inspect           
typing_extensions        
tzdata                   
unicodedata2             
urllib3                  
vc                       
visdom                   
vs2015_runtime           
wcwidth                  
websocket-client         
werkzeug                 
wheel                    
win_inet_pton            
xz                       
yaml                     
zeromq                   
zipp                     
zlib                     
zstd                     